Fix PlayerInputState ambiguity after refactor
